Module contract_bindings::hot_shot::hot_shot

This module was auto-generated with ethers-rs Abigen. More information at: https://github.com/gakonst/ethers-rs

Structs

AddNewStakingKeyCall

Container type for all input parameters for the addNewStakingKey function with signature addNewStakingKey((uint256,uint256,uint256,uint256),uint256) and selector 0xf1f45d99

BlockHeightCall

Container type for all input parameters for the blockHeight function with signature blockHeight() and selector 0xf44ff712

BlockHeightReturn

Container type for all return fields from the blockHeight function with signature blockHeight() and selector 0xf44ff712

CommitmentsCall

Container type for all input parameters for the commitments function with signature commitments(uint256) and selector 0x49ce8997

CommitmentsReturn

Container type for all return fields from the commitments function with signature commitments(uint256) and selector 0x49ce8997

G2Point

G2Point(uint256,uint256,uint256,uint256)

GetStakingKeyCall

Container type for all input parameters for the getStakingKey function with signature getStakingKey(uint256) and selector 0x67a21e70

GetStakingKeyReturn

Container type for all return fields from the getStakingKey function with signature getStakingKey(uint256) and selector 0x67a21e70

HotShot

IncorrectBlockNumber

Custom Error type IncorrectBlockNumber with signature IncorrectBlockNumber(uint256,uint256) and selector 0x34e423ff

InvalidQC

Custom Error type InvalidQC with signature InvalidQC(uint256) and selector 0x78186719

MaxBlocksCall

Container type for all input parameters for the MAX_BLOCKS function with signature MAX_BLOCKS() and selector 0x26833dcc

MaxBlocksReturn

Container type for all return fields from the MAX_BLOCKS function with signature MAX_BLOCKS() and selector 0x26833dcc

NewBlocksCall

Container type for all input parameters for the newBlocks function with signature newBlocks((uint256,uint256,uint256,uint256)[]) and selector 0x0a321cff

NewBlocksFilter

NewStakingKeyFilter

NoKeySelected

Custom Error type NoKeySelected with signature NoKeySelected() and selector 0x96cb045a

NotEnoughStake

Custom Error type NotEnoughStake with signature NotEnoughStake() and selector 0xf0a42d4c

Qc

Qc(uint256,uint256,uint256,uint256)

TooManyBlocks

Custom Error type TooManyBlocks with signature TooManyBlocks(uint256) and selector 0xe082840b

Enums

HotShotCalls

Container type for all of the contract’s call

HotShotErrors

Container type for all of the contract’s custom errors

HotShotEvents

Container type for all of the contract’s events

Constants

__BYTECODE 🔒

__DEPLOYED_BYTECODE 🔒

Statics

HOTSHOT_ABI

The parsed JSON ABI of the contract.

HOTSHOT_BYTECODE

The bytecode of the contract.

HOTSHOT_DEPLOYED_BYTECODE

The deployed bytecode of the contract.

Functions

__abi 🔒