Module contract_bindings::light_client::light_client

This module was auto-generated with ethers-rs Abigen. More information at: https://github.com/gakonst/ethers-rs

Re-exports

pub use super::super::shared_types::*;

Structs

AddressEmptyCode

Custom Error type AddressEmptyCode with signature AddressEmptyCode(address) and selector 0x9996b315

BlocksPerEpochCall

Container type for all input parameters for the blocksPerEpoch function with signature blocksPerEpoch() and selector 0xf0682054

BlocksPerEpochReturn

Container type for all return fields from the blocksPerEpoch function with signature blocksPerEpoch() and selector 0xf0682054

ComputeStakeTableCommCall

Container type for all input parameters for the computeStakeTableComm function with signature computeStakeTableComm((uint64,uint64,uint256,uint256,uint256,uint256,uint256,uint256)) and selector 0xaa922732

ComputeStakeTableCommReturn

Container type for all return fields from the computeStakeTableComm function with signature computeStakeTableComm((uint64,uint64,uint256,uint256,uint256,uint256,uint256,uint256)) and selector 0xaa922732

CurrentEpochCall

Container type for all input parameters for the currentEpoch function with signature currentEpoch() and selector 0x76671808

CurrentEpochReturn

Container type for all return fields from the currentEpoch function with signature currentEpoch() and selector 0x76671808

DisablePermissionedProverModeCall

Container type for all input parameters for the disablePermissionedProverMode function with signature disablePermissionedProverMode() and selector 0x69cc6a04

ERC1967InvalidImplementation

Custom Error type ERC1967InvalidImplementation with signature ERC1967InvalidImplementation(address) and selector 0x4c9c8ce3

ERC1967NonPayable

Custom Error type ERC1967NonPayable with signature ERC1967NonPayable() and selector 0xb398979f

EpochChangedFilter

FailedInnerCall

Custom Error type FailedInnerCall with signature FailedInnerCall() and selector 0x1425ea42

FrozenStakeTableCommitmentCall

Container type for all input parameters for the frozenStakeTableCommitment function with signature frozenStakeTableCommitment() and selector 0x382b215a

FrozenStakeTableCommitmentReturn

Container type for all return fields from the frozenStakeTableCommitment function with signature frozenStakeTableCommitment() and selector 0x382b215a

FrozenThresholdCall

Container type for all input parameters for the frozenThreshold function with signature frozenThreshold() and selector 0xca6fe855

FrozenThresholdReturn

Container type for all return fields from the frozenThreshold function with signature frozenThreshold() and selector 0xca6fe855

GetFinalizedStateCall

Container type for all input parameters for the getFinalizedState function with signature getFinalizedState() and selector 0x82d07ff3

GetFinalizedStateReturn

Container type for all return fields from the getFinalizedState function with signature getFinalizedState() and selector 0x82d07ff3

GetGenesisStateCall

Container type for all input parameters for the getGenesisState function with signature getGenesisState() and selector 0x4847ae5d

GetGenesisStateReturn

Container type for all return fields from the getGenesisState function with signature getGenesisState() and selector 0x4847ae5d

GetVersionCall

Container type for all input parameters for the getVersion function with signature getVersion() and selector 0x0d8e6e2c

GetVersionReturn

Container type for all return fields from the getVersion function with signature getVersion() and selector 0x0d8e6e2c

InitializeCall

Container type for all input parameters for the initialize function with signature initialize((uint64,uint64,uint256,uint256,uint256,uint256,uint256,uint256),uint32,address) and selector 0xa244d596

InitializedFilter

InvalidAddress

Custom Error type InvalidAddress with signature InvalidAddress() and selector 0xe6c4247b

InvalidArgs

Custom Error type InvalidArgs with signature InvalidArgs() and selector 0xa1ba07ee

InvalidInitialization

Custom Error type InvalidInitialization with signature InvalidInitialization() and selector 0xf92ee8a9

InvalidPolyEvalArgs

Custom Error type InvalidPolyEvalArgs with signature InvalidPolyEvalArgs() and selector 0x8c5e11f1

InvalidProof

Custom Error type InvalidProof with signature InvalidProof() and selector 0x09bde339

LightClient

MissingLastBlockForCurrentEpoch

Custom Error type MissingLastBlockForCurrentEpoch with signature MissingLastBlockForCurrentEpoch(uint64) and selector 0x1b2335f8

NewFinalizedStateCall

Container type for all input parameters for the newFinalizedState function with signature newFinalizedState((uint64,uint64,uint256,uint256,uint256,uint256,uint256,uint256),((uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256)) and selector 0x409939b7

NewStateFilter

NoChangeRequired

Custom Error type NoChangeRequired with signature NoChangeRequired() and selector 0xa863aec9

NotInitializing

Custom Error type NotInitializing with signature NotInitializing() and selector 0xd7e6bcf8

OutdatedState

Custom Error type OutdatedState with signature OutdatedState() and selector 0x051c46ef

OwnableInvalidOwner

Custom Error type OwnableInvalidOwner with signature OwnableInvalidOwner(address) and selector 0x1e4fbdf7

OwnableUnauthorizedAccount

Custom Error type OwnableUnauthorizedAccount with signature OwnableUnauthorizedAccount(address) and selector 0x118cdaa7

OwnerCall

Container type for all input parameters for the owner function with signature owner() and selector 0x8da5cb5b

OwnerReturn

Container type for all return fields from the owner function with signature owner() and selector 0x8da5cb5b

OwnershipTransferredFilter

PermissionedProverCall

Container type for all input parameters for the permissionedProver function with signature permissionedProver() and selector 0x313df7b1

PermissionedProverEnabledCall

Container type for all input parameters for the permissionedProverEnabled function with signature permissionedProverEnabled() and selector 0xbd32519a

PermissionedProverEnabledReturn

Container type for all return fields from the permissionedProverEnabled function with signature permissionedProverEnabled() and selector 0xbd32519a

PermissionedProverNotRequiredFilter

PermissionedProverNotSet

Custom Error type PermissionedProverNotSet with signature PermissionedProverNotSet() and selector 0x25cda3ce

PermissionedProverRequiredFilter

PermissionedProverReturn

Container type for all return fields from the permissionedProver function with signature permissionedProver() and selector 0x313df7b1

ProverNotPermissioned

Custom Error type ProverNotPermissioned with signature ProverNotPermissioned() and selector 0xa3a64780

ProxiableUUIDCall

Container type for all input parameters for the proxiableUUID function with signature proxiableUUID() and selector 0x52d1902d

ProxiableUUIDReturn

Container type for all return fields from the proxiableUUID function with signature proxiableUUID() and selector 0x52d1902d

RenounceOwnershipCall

Container type for all input parameters for the renounceOwnership function with signature renounceOwnership() and selector 0x715018a6

SetPermissionedProverCall

Container type for all input parameters for the setPermissionedProver function with signature setPermissionedProver(address) and selector 0x013fa5fc

StatesCall

Container type for all input parameters for the states function with signature states(uint32) and selector 0x7f17baad

StatesReturn

Container type for all return fields from the states function with signature states(uint32) and selector 0x7f17baad

TransferOwnershipCall

Container type for all input parameters for the transferOwnership function with signature transferOwnership(address) and selector 0xf2fde38b

UUPSUnauthorizedCallContext

Custom Error type UUPSUnauthorizedCallContext with signature UUPSUnauthorizedCallContext() and selector 0xe07c8dba

UUPSUnsupportedProxiableUUID

Custom Error type UUPSUnsupportedProxiableUUID with signature UUPSUnsupportedProxiableUUID(bytes32) and selector 0xaa1d49a4

UnsupportedDegree

Custom Error type UnsupportedDegree with signature UnsupportedDegree() and selector 0xe2ef09e5

UpgradeFilter

UpgradeInterfaceVersionCall

Container type for all input parameters for the UPGRADE_INTERFACE_VERSION function with signature UPGRADE_INTERFACE_VERSION() and selector 0xad3cb1cc

UpgradeInterfaceVersionReturn

Container type for all return fields from the UPGRADE_INTERFACE_VERSION function with signature UPGRADE_INTERFACE_VERSION() and selector 0xad3cb1cc

UpgradeToAndCallCall

Container type for all input parameters for the upgradeToAndCall function with signature upgradeToAndCall(address,bytes) and selector 0x4f1ef286

UpgradedFilter

VotingStakeTableCommitmentCall

Container type for all input parameters for the votingStakeTableCommitment function with signature votingStakeTableCommitment() and selector 0x76b6b7cb

VotingStakeTableCommitmentReturn

Container type for all return fields from the votingStakeTableCommitment function with signature votingStakeTableCommitment() and selector 0x76b6b7cb

VotingThresholdCall

Container type for all input parameters for the votingThreshold function with signature votingThreshold() and selector 0x62827733

VotingThresholdReturn

Container type for all return fields from the votingThreshold function with signature votingThreshold() and selector 0x62827733

WrongPlonkVK

Custom Error type WrongPlonkVK with signature WrongPlonkVK() and selector 0x41f53b12

WrongStakeTableUsed

Custom Error type WrongStakeTableUsed with signature WrongStakeTableUsed() and selector 0x51618089

Enums

LightClientCalls

Container type for all of the contract’s call

LightClientErrors

Container type for all of the contract’s custom errors

LightClientEvents

Container type for all of the contract’s events

Constants

__BYTECODE 🔒

__DEPLOYED_BYTECODE 🔒

Statics

LIGHTCLIENT_ABI

The parsed JSON ABI of the contract.

LIGHTCLIENT_BYTECODE

The bytecode of the contract.

LIGHTCLIENT_DEPLOYED_BYTECODE

The deployed bytecode of the contract.

Functions

__abi 🔒