Module contract_bindings::light_client_mock::light_client_mock

This module was auto-generated with ethers-rs Abigen. More information at: https://github.com/gakonst/ethers-rs

Re-exports

pub use super::super::shared_types::*;

Structs

AddressEmptyCode

Custom Error type AddressEmptyCode with signature AddressEmptyCode(address) and selector 0x9996b315

ComputeStakeTableCommCall

Container type for all input parameters for the computeStakeTableComm function with signature computeStakeTableComm((uint64,uint64,uint256,uint256,uint256,uint256,uint256,uint256)) and selector 0xaa922732

ComputeStakeTableCommReturn

Container type for all return fields from the computeStakeTableComm function with signature computeStakeTableComm((uint64,uint64,uint256,uint256,uint256,uint256,uint256,uint256)) and selector 0xaa922732

DisablePermissionedProverModeCall

Container type for all input parameters for the disablePermissionedProverMode function with signature disablePermissionedProverMode() and selector 0x69cc6a04

ERC1967InvalidImplementation

Custom Error type ERC1967InvalidImplementation with signature ERC1967InvalidImplementation(address) and selector 0x4c9c8ce3

ERC1967NonPayable

Custom Error type ERC1967NonPayable with signature ERC1967NonPayable() and selector 0xb398979f

FailedInnerCall

Custom Error type FailedInnerCall with signature FailedInnerCall() and selector 0x1425ea42

FrozenStakeTableCommitmentCall

Container type for all input parameters for the frozenStakeTableCommitment function with signature frozenStakeTableCommitment() and selector 0x382b215a

FrozenStakeTableCommitmentReturn

Container type for all return fields from the frozenStakeTableCommitment function with signature frozenStakeTableCommitment() and selector 0x382b215a

FrozenThresholdCall

Container type for all input parameters for the frozenThreshold function with signature frozenThreshold() and selector 0xca6fe855

FrozenThresholdReturn

Container type for all return fields from the frozenThreshold function with signature frozenThreshold() and selector 0xca6fe855

GetFinalizedStateCall

Container type for all input parameters for the getFinalizedState function with signature getFinalizedState() and selector 0x82d07ff3

GetFinalizedStateReturn

Container type for all return fields from the getFinalizedState function with signature getFinalizedState() and selector 0x82d07ff3

GetGenesisStateCall

Container type for all input parameters for the getGenesisState function with signature getGenesisState() and selector 0x4847ae5d

GetGenesisStateReturn

Container type for all return fields from the getGenesisState function with signature getGenesisState() and selector 0x4847ae5d

GetHotShotCommitmentCall

Container type for all input parameters for the getHotShotCommitment function with signature getHotShotCommitment(uint256) and selector 0x8584d23f

GetHotShotCommitmentReturn

Container type for all return fields from the getHotShotCommitment function with signature getHotShotCommitment(uint256) and selector 0x8584d23f

GetStateHistoryCountCall

Container type for all input parameters for the getStateHistoryCount function with signature getStateHistoryCount() and selector 0xf9e50d19

GetStateHistoryCountReturn

Container type for all return fields from the getStateHistoryCount function with signature getStateHistoryCount() and selector 0xf9e50d19

GetVersionCall

Container type for all input parameters for the getVersion function with signature getVersion() and selector 0x0d8e6e2c

GetVersionReturn

Container type for all return fields from the getVersion function with signature getVersion() and selector 0x0d8e6e2c

InitializeCall

Container type for all input parameters for the initialize function with signature initialize((uint64,uint64,uint256,uint256,uint256,uint256,uint256,uint256),uint32,address) and selector 0xa244d596

InitializedFilter

InsufficientSnapshotHistory

Custom Error type InsufficientSnapshotHistory with signature InsufficientSnapshotHistory() and selector 0xb0b43877

InvalidAddress

Custom Error type InvalidAddress with signature InvalidAddress() and selector 0xe6c4247b

InvalidArgs

Custom Error type InvalidArgs with signature InvalidArgs() and selector 0xa1ba07ee

InvalidHotShotBlockForCommitmentCheck

Custom Error type InvalidHotShotBlockForCommitmentCheck with signature InvalidHotShotBlockForCommitmentCheck() and selector 0x615a9264

InvalidInitialization

Custom Error type InvalidInitialization with signature InvalidInitialization() and selector 0xf92ee8a9

InvalidMaxStateHistory

Custom Error type InvalidMaxStateHistory with signature InvalidMaxStateHistory() and selector 0xf4a0eee0

InvalidProof

Custom Error type InvalidProof with signature InvalidProof() and selector 0x09bde339

LagOverEscapeHatchThresholdCall

Container type for all input parameters for the lagOverEscapeHatchThreshold function with signature lagOverEscapeHatchThreshold(uint256,uint256) and selector 0xe0303301

LagOverEscapeHatchThresholdReturn

Container type for all return fields from the lagOverEscapeHatchThreshold function with signature lagOverEscapeHatchThreshold(uint256,uint256) and selector 0xe0303301

LightClientMock

NewFinalizedStateCall

Container type for all input parameters for the newFinalizedState function with signature newFinalizedState((uint64,uint64,uint256,uint256,uint256,uint256,uint256,uint256),((uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256)) and selector 0x409939b7

NewStateFilter

NoChangeRequired

Custom Error type NoChangeRequired with signature NoChangeRequired() and selector 0xa863aec9

NotInitializing

Custom Error type NotInitializing with signature NotInitializing() and selector 0xd7e6bcf8

OutdatedState

Custom Error type OutdatedState with signature OutdatedState() and selector 0x051c46ef

OwnableInvalidOwner

Custom Error type OwnableInvalidOwner with signature OwnableInvalidOwner(address) and selector 0x1e4fbdf7

OwnableUnauthorizedAccount

Custom Error type OwnableUnauthorizedAccount with signature OwnableUnauthorizedAccount(address) and selector 0x118cdaa7

OwnerCall

Container type for all input parameters for the owner function with signature owner() and selector 0x8da5cb5b

OwnerReturn

Container type for all return fields from the owner function with signature owner() and selector 0x8da5cb5b

OwnershipTransferredFilter

PermissionedProverCall

Container type for all input parameters for the permissionedProver function with signature permissionedProver() and selector 0x313df7b1

PermissionedProverEnabledCall

Container type for all input parameters for the permissionedProverEnabled function with signature permissionedProverEnabled() and selector 0xbd32519a

PermissionedProverEnabledReturn

Container type for all return fields from the permissionedProverEnabled function with signature permissionedProverEnabled() and selector 0xbd32519a

PermissionedProverNotRequiredFilter

PermissionedProverRequiredFilter

PermissionedProverReturn

Container type for all return fields from the permissionedProver function with signature permissionedProver() and selector 0x313df7b1

ProverNotPermissioned

Custom Error type ProverNotPermissioned with signature ProverNotPermissioned() and selector 0xa3a64780

ProxiableUUIDCall

Container type for all input parameters for the proxiableUUID function with signature proxiableUUID() and selector 0x52d1902d

ProxiableUUIDReturn

Container type for all return fields from the proxiableUUID function with signature proxiableUUID() and selector 0x52d1902d

RenounceOwnershipCall

Container type for all input parameters for the renounceOwnership function with signature renounceOwnership() and selector 0x715018a6

SetFinalizedStateCall

Container type for all input parameters for the setFinalizedState function with signature setFinalizedState((uint64,uint64,uint256,uint256,uint256,uint256,uint256,uint256)) and selector 0x202a0adb

SetHotShotDownSinceCall

Container type for all input parameters for the setHotShotDownSince function with signature setHotShotDownSince(uint256) and selector 0x2d52aad6

SetHotShotUpCall

Container type for all input parameters for the setHotShotUp function with signature setHotShotUp() and selector 0xc8e5e498

SetPermissionedProverCall

Container type for all input parameters for the setPermissionedProver function with signature setPermissionedProver(address) and selector 0x013fa5fc

SetStateHistoryCall

Container type for all input parameters for the setStateHistory function with signature setStateHistory((uint64,uint64,(uint64,uint256))[]) and selector 0x2f5f4600

SetstateHistoryRetentionPeriodCall

Container type for all input parameters for the setstateHistoryRetentionPeriod function with signature setstateHistoryRetentionPeriod(uint32) and selector 0x96c1ca61

StateHistoryCommitment

StateHistoryCommitment(uint64,uint64,(uint64,uint256))

StateHistoryCommitmentsCall

Container type for all input parameters for the stateHistoryCommitments function with signature stateHistoryCommitments(uint256) and selector 0x02b592f3

StateHistoryCommitmentsReturn

Container type for all return fields from the stateHistoryCommitments function with signature stateHistoryCommitments(uint256) and selector 0x02b592f3

StateHistoryFirstIndexCall

Container type for all input parameters for the stateHistoryFirstIndex function with signature stateHistoryFirstIndex() and selector 0x2f79889d

StateHistoryFirstIndexReturn

Container type for all return fields from the stateHistoryFirstIndex function with signature stateHistoryFirstIndex() and selector 0x2f79889d

StateHistoryRetentionPeriodCall

Container type for all input parameters for the stateHistoryRetentionPeriod function with signature stateHistoryRetentionPeriod() and selector 0xc23b9e9e

StateHistoryRetentionPeriodReturn

Container type for all return fields from the stateHistoryRetentionPeriod function with signature stateHistoryRetentionPeriod() and selector 0xc23b9e9e

StatesCall

Container type for all input parameters for the states function with signature states(uint32) and selector 0x7f17baad

StatesReturn

Container type for all return fields from the states function with signature states(uint32) and selector 0x7f17baad

TransferOwnershipCall

Container type for all input parameters for the transferOwnership function with signature transferOwnership(address) and selector 0xf2fde38b

UUPSUnauthorizedCallContext

Custom Error type UUPSUnauthorizedCallContext with signature UUPSUnauthorizedCallContext() and selector 0xe07c8dba

UUPSUnsupportedProxiableUUID

Custom Error type UUPSUnsupportedProxiableUUID with signature UUPSUnsupportedProxiableUUID(bytes32) and selector 0xaa1d49a4

UpgradeFilter

UpgradeInterfaceVersionCall

Container type for all input parameters for the UPGRADE_INTERFACE_VERSION function with signature UPGRADE_INTERFACE_VERSION() and selector 0xad3cb1cc

UpgradeInterfaceVersionReturn

Container type for all return fields from the UPGRADE_INTERFACE_VERSION function with signature UPGRADE_INTERFACE_VERSION() and selector 0xad3cb1cc

UpgradeToAndCallCall

Container type for all input parameters for the upgradeToAndCall function with signature upgradeToAndCall(address,bytes) and selector 0x4f1ef286

UpgradedFilter

VotingStakeTableCommitmentCall

Container type for all input parameters for the votingStakeTableCommitment function with signature votingStakeTableCommitment() and selector 0x76b6b7cb

VotingStakeTableCommitmentReturn

Container type for all return fields from the votingStakeTableCommitment function with signature votingStakeTableCommitment() and selector 0x76b6b7cb

VotingThresholdCall

Container type for all input parameters for the votingThreshold function with signature votingThreshold() and selector 0x62827733

VotingThresholdReturn

Container type for all return fields from the votingThreshold function with signature votingThreshold() and selector 0x62827733

WrongStakeTableUsed

Custom Error type WrongStakeTableUsed with signature WrongStakeTableUsed() and selector 0x51618089

Enums

LightClientMockCalls

Container type for all of the contract’s call

LightClientMockErrors

Container type for all of the contract’s custom errors

LightClientMockEvents

Container type for all of the contract’s events

Statics

LIGHTCLIENTMOCK_ABI

The parsed JSON ABI of the contract.

Functions

__abi 🔒