contract_bindings::permissioned_stake_table
Module permissioned_stake_table

This module was auto-generated with ethers-rs Abigen. More information at: https://github.com/gakonst/ethers-rs

Structs

EdOnBN254Point

EdOnBN254Point(uint256,uint256)

G2Point

G2Point(uint256,uint256,uint256,uint256)

HashBlsKeyCall

Container type for all input parameters for the _hashBlsKey function with signature _hashBlsKey((uint256,uint256,uint256,uint256)) and selector 0x9b30a5e6

HashBlsKeyReturn

Container type for all return fields from the _hashBlsKey function with signature _hashBlsKey((uint256,uint256,uint256,uint256)) and selector 0x9b30a5e6

IsStakerCall

Container type for all input parameters for the isStaker function with signature isStaker((uint256,uint256,uint256,uint256)) and selector 0x75d705e9

IsStakerReturn

Container type for all return fields from the isStaker function with signature isStaker((uint256,uint256,uint256,uint256)) and selector 0x75d705e9

NodeInfo

NodeInfo((uint256,uint256,uint256,uint256),(uint256,uint256),bool)

OwnableInvalidOwner

Custom Error type OwnableInvalidOwner with signature OwnableInvalidOwner(address) and selector 0x1e4fbdf7

OwnableUnauthorizedAccount

Custom Error type OwnableUnauthorizedAccount with signature OwnableUnauthorizedAccount(address) and selector 0x118cdaa7

OwnerCall

Container type for all input parameters for the owner function with signature owner() and selector 0x8da5cb5b

OwnerReturn

Container type for all return fields from the owner function with signature owner() and selector 0x8da5cb5b

OwnershipTransferredFilter

PermissionedStakeTable

RenounceOwnershipCall

Container type for all input parameters for the renounceOwnership function with signature renounceOwnership() and selector 0x715018a6

StakerAlreadyExists

Custom Error type StakerAlreadyExists with signature StakerAlreadyExists((uint256,uint256,uint256,uint256)) and selector 0x360dc282

StakerNotFound

Custom Error type StakerNotFound with signature StakerNotFound((uint256,uint256,uint256,uint256)) and selector 0x34a7561f

StakersUpdatedFilter

TransferOwnershipCall

Container type for all input parameters for the transferOwnership function with signature transferOwnership(address) and selector 0xf2fde38b

UpdateCall

Container type for all input parameters for the update function with signature update(((uint256,uint256,uint256,uint256),(uint256,uint256),bool)[],((uint256,uint256,uint256,uint256),(uint256,uint256),bool)[]) and selector 0xa8a0ea5c

Enums

PermissionedStakeTableCalls

Container type for all of the contract’s call

PermissionedStakeTableErrors

Container type for all of the contract’s custom errors

PermissionedStakeTableEvents

Container type for all of the contract’s events

Constants

__BYTECODE 🔒

__DEPLOYED_BYTECODE 🔒

Statics

PERMISSIONEDSTAKETABLE_ABI

The parsed JSON ABI of the contract.

PERMISSIONEDSTAKETABLE_BYTECODE

The bytecode of the contract.

PERMISSIONEDSTAKETABLE_DEPLOYED_BYTECODE

The deployed bytecode of the contract.

Functions

__abi 🔒