Module contract_bindings::plonk_verifier::plonk_verifier

This module was auto-generated with ethers-rs Abigen. More information at: https://github.com/gakonst/ethers-rs

Re-exports

pub use super::super::shared_types::*;

Structs

InvalidPlonkArgs

Custom Error type InvalidPlonkArgs with signature InvalidPlonkArgs() and selector 0xfd9a2d1b

InvalidPolyEvalArgs

Custom Error type InvalidPolyEvalArgs with signature InvalidPolyEvalArgs() and selector 0x8c5e11f1

PlonkVerifier

UnsupportedDegree

Custom Error type UnsupportedDegree with signature UnsupportedDegree() and selector 0xe2ef09e5

VerifyCall

Container type for all input parameters for the verify function with signature verify((uint256,uint256,(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256)),uint256[],((uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256)) and selector 0x8c266f05

VerifyReturn

Container type for all return fields from the verify function with signature verify((uint256,uint256,(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256)),uint256[],((uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256)) and selector 0x8c266f05

WrongPlonkVK

Custom Error type WrongPlonkVK with signature WrongPlonkVK() and selector 0x41f53b12

Enums

PlonkVerifierErrors

Container type for all of the contract’s custom errors

Constants

__BYTECODE 🔒

__DEPLOYED_BYTECODE 🔒

Statics

PLONKVERIFIER_ABI

The parsed JSON ABI of the contract.

PLONKVERIFIER_BYTECODE

The bytecode of the contract.

PLONKVERIFIER_DEPLOYED_BYTECODE

The deployed bytecode of the contract.

Functions

__abi 🔒