Module contract_bindings::fee_contract::fee_contract

This module was auto-generated with ethers-rs Abigen. More information at: https://github.com/gakonst/ethers-rs

Structs

AddressEmptyCode

Custom Error type AddressEmptyCode with signature AddressEmptyCode(address) and selector 0x9996b315

BalancesCall

Container type for all input parameters for the balances function with signature balances(address) and selector 0x27e235e3

BalancesReturn

Container type for all return fields from the balances function with signature balances(address) and selector 0x27e235e3

DepositCall

Container type for all input parameters for the deposit function with signature deposit(address) and selector 0xf340fa01

DepositFilter

DepositTooLarge

Custom Error type DepositTooLarge with signature DepositTooLarge() and selector 0xc56d46d3

DepositTooSmall

Custom Error type DepositTooSmall with signature DepositTooSmall() and selector 0x6ba4a1c7

ERC1967InvalidImplementation

Custom Error type ERC1967InvalidImplementation with signature ERC1967InvalidImplementation(address) and selector 0x4c9c8ce3

ERC1967NonPayable

Custom Error type ERC1967NonPayable with signature ERC1967NonPayable() and selector 0xb398979f

FailedInnerCall

Custom Error type FailedInnerCall with signature FailedInnerCall() and selector 0x1425ea42

FeeContract

FunctionDoesNotExist

Custom Error type FunctionDoesNotExist with signature FunctionDoesNotExist() and selector 0xa9ad62f8

GetVersionCall

Container type for all input parameters for the getVersion function with signature getVersion() and selector 0x0d8e6e2c

GetVersionReturn

Container type for all return fields from the getVersion function with signature getVersion() and selector 0x0d8e6e2c

InitializeCall

Container type for all input parameters for the initialize function with signature initialize(address) and selector 0xc4d66de8

InitializedFilter

InvalidInitialization

Custom Error type InvalidInitialization with signature InvalidInitialization() and selector 0xf92ee8a9

InvalidUserAddress

Custom Error type InvalidUserAddress with signature InvalidUserAddress() and selector 0x702b3d90

LogFilter

MaxDepositAmountCall

Container type for all input parameters for the maxDepositAmount function with signature maxDepositAmount() and selector 0x8ed83271

MaxDepositAmountReturn

Container type for all return fields from the maxDepositAmount function with signature maxDepositAmount() and selector 0x8ed83271

MinDepositAmountCall

Container type for all input parameters for the minDepositAmount function with signature minDepositAmount() and selector 0x645006ca

MinDepositAmountReturn

Container type for all return fields from the minDepositAmount function with signature minDepositAmount() and selector 0x645006ca

NoFunctionCalled

Custom Error type NoFunctionCalled with signature NoFunctionCalled() and selector 0xbc8eca1b

NotInitializing

Custom Error type NotInitializing with signature NotInitializing() and selector 0xd7e6bcf8

OwnableInvalidOwner

Custom Error type OwnableInvalidOwner with signature OwnableInvalidOwner(address) and selector 0x1e4fbdf7

OwnableUnauthorizedAccount

Custom Error type OwnableUnauthorizedAccount with signature OwnableUnauthorizedAccount(address) and selector 0x118cdaa7

OwnerCall

Container type for all input parameters for the owner function with signature owner() and selector 0x8da5cb5b

OwnerReturn

Container type for all return fields from the owner function with signature owner() and selector 0x8da5cb5b

OwnershipTransferredFilter

ProxiableUUIDCall

Container type for all input parameters for the proxiableUUID function with signature proxiableUUID() and selector 0x52d1902d

ProxiableUUIDReturn

Container type for all return fields from the proxiableUUID function with signature proxiableUUID() and selector 0x52d1902d

RenounceOwnershipCall

Container type for all input parameters for the renounceOwnership function with signature renounceOwnership() and selector 0x715018a6

TransferOwnershipCall

Container type for all input parameters for the transferOwnership function with signature transferOwnership(address) and selector 0xf2fde38b

UUPSUnauthorizedCallContext

Custom Error type UUPSUnauthorizedCallContext with signature UUPSUnauthorizedCallContext() and selector 0xe07c8dba

UUPSUnsupportedProxiableUUID

Custom Error type UUPSUnsupportedProxiableUUID with signature UUPSUnsupportedProxiableUUID(bytes32) and selector 0xaa1d49a4

UpgradeFilter

UpgradeInterfaceVersionCall

Container type for all input parameters for the UPGRADE_INTERFACE_VERSION function with signature UPGRADE_INTERFACE_VERSION() and selector 0xad3cb1cc

UpgradeInterfaceVersionReturn

Container type for all return fields from the UPGRADE_INTERFACE_VERSION function with signature UPGRADE_INTERFACE_VERSION() and selector 0xad3cb1cc

UpgradeToAndCallCall

Container type for all input parameters for the upgradeToAndCall function with signature upgradeToAndCall(address,bytes) and selector 0x4f1ef286

UpgradedFilter

Enums

FeeContractCalls

Container type for all of the contract’s call

FeeContractErrors

Container type for all of the contract’s custom errors

FeeContractEvents

Container type for all of the contract’s events

Constants

__BYTECODE 🔒

__DEPLOYED_BYTECODE 🔒

Statics

FEECONTRACT_ABI

The parsed JSON ABI of the contract.

FEECONTRACT_BYTECODE

The bytecode of the contract.

FEECONTRACT_DEPLOYED_BYTECODE

The deployed bytecode of the contract.

Functions

__abi 🔒