contract_bindings::plonk_verifier_2
Module plonk_verifier_2

This module was auto-generated with ethers-rs Abigen. More information at: https://github.com/gakonst/ethers-rs

Re-exports

pub use super::super::shared_types::*;

Structs

PModCall

Container type for all input parameters for the P_MOD function with signature P_MOD() and selector 0x1d712e27

PModReturn

Container type for all return fields from the P_MOD function with signature P_MOD() and selector 0x1d712e27

PlonkVerifier2

RModCall

Container type for all input parameters for the R_MOD function with signature R_MOD() and selector 0xdf6e6cb4

RModReturn

Container type for all return fields from the R_MOD function with signature R_MOD() and selector 0xdf6e6cb4

UnsupportedDegree

Custom Error type UnsupportedDegree with signature UnsupportedDegree() and selector 0xe2ef09e5

VerifyCall

Container type for all input parameters for the verify function with signature verify((uint256,uint256,(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),bytes32,bytes32),uint256[7],((uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256)) and selector 0x77ae3f07

VerifyReturn

Container type for all return fields from the verify function with signature verify((uint256,uint256,(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),bytes32,bytes32),uint256[7],((uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256)) and selector 0x77ae3f07

Enums

PlonkVerifier2Calls

Container type for all of the contract’s call

Constants

__BYTECODE 🔒

__DEPLOYED_BYTECODE 🔒

Statics

PLONKVERIFIER2_ABI

The parsed JSON ABI of the contract.

PLONKVERIFIER2_BYTECODE

The bytecode of the contract.

PLONKVERIFIER2_DEPLOYED_BYTECODE

The deployed bytecode of the contract.

Functions

__abi 🔒